todas las vistas siguen conectadas y ahora esa funcional la opcion de crear guias rapidas, falta poder editarlas y hacer lo mismo con los otros objetos

+++ src/main/java/com/LityAppAdmin/Repository/IGuiaRapidaRepository.java
@@ -3,5 +3,5 @@ package com.LityAppAdmin.Repository;
 import com.LityAppAdmin.Model.GuiaRapidaModel;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IGuiaRapidaRepository extends JpaRepository<GuiaRapidaModel, Integer> {
+public interface IGuiaRapidaRepository extends JpaRepository<GuiaRapidaModel, Long> {
 }
